Split user tests into their own file
